configs: rcar-gen3: Use the standard distro boot commands
The standard hardcoded tftp boot setup will not be very useful for most
people. U-boot has had a quite flexible standard boot sequence for a
while now (see doc/README.distro) so lets use that so it's easier to
make images that just work out of the box on these devices.

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>